+++ bin/citrea/src/rollup/bitcoin.rs
@@ -10,8 +10,7 @@ use citrea_common::{FullNodeConfig, ProverConfig};
 use citrea_primitives::{REVEAL_BATCH_PROOF_PREFIX, REVEAL_LIGHT_CLIENT_PREFIX};
 use citrea_prover::prover_service::ParallelProverService;
 // use citrea_risc0_bonsai_adapter::host::Risc0BonsaiHost;
-use citrea_sp1::host::SP1Host;
-use citrea_sp1::VerifyingKey;
+use citrea_sp1::host::{SP1Host, VerifyingKey};
 use citrea_stf::genesis_config::StorageConfig;
 use citrea_stf::runtime::Runtime;
 use sov_db::ledger_db::LedgerDB;
