+++ crates/sp1/src/guest.rs
@@ -3,8 +3,6 @@ use borsh::{BorshDeserialize, BorshSerialize};
 use sov_rollup_interface::zk::{Zkvm, ZkvmGuest};
 use sp1_zkvm::io;
 
-use crate::VerifyingKey;
-
 /// A guest for the SP1 VM. Implements the `ZkvmGuest` trait
 ///  in terms of Risc0's env::read and env::commit functions.
 #[derive(Default)]
@@ -18,7 +16,11 @@ impl SP1Guest {
 }
 
 impl Zkvm for SP1Guest {
-    type CodeCommitment = VerifyingKey;
+    #[cfg(feature = "native")]
+    type CodeCommitment = crate::host::VerifyingKey;
+    #[cfg(not(feature = "native"))]
+    type CodeCommitment = ();
+
     type Error = anyhow::Error;
 
     fn verify(
